# HEIGVD-AMT-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  title: Gamification API
  description: Move your app forward with the RFPS API.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
paths:
...
  /applications:
    get:
      description: 
        The Applications endpoint returns information about the application whose id is present in the authentication token. The response includes the application id, the application name and the application description.
      produces:
        - application/json
      tags:
        - Applications
      parameters:
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        200:
          description: OK, returns the application.
          schema:
              $ref: '#/definitions/ApplicationOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.
        500:
          description: Internal server error, if database isn't reachable.
          
    post:
      description: Create a new application.
      consumes:
        - application/json
      parameters:
        - in: body
          name: application
          description: Application object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/ApplicationInputDTO'
      responses:
        201:
          description: Created, if it saved successfully.
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name or password fields contain more than 80 characters,
                       if description field contain more than 255 characters,
                       if password contains less than 7 characters,
                       if application name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.
          
    put:
      description: Modify completely this application whose id is present in the authentication token.
      consumes:
        - application/json
      parameters:
        - name: application
          in: body
          description: Application to be updated
          required: true
          schema:
              $ref: '#/definitions/ApplicationInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name or password fields contain more than 80 characters,
                       if description field contain more than 255 characters,
                       if password contains less than 7 characters,
                       if application name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.

    delete:
      description: Delete the application which has the specified id.
      parameters:
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        410:
          description: Gone, if authentication token contains an application_id that doesn't exist. Typically, if the application was deleted.
        500:
          description: Internal server error, if database isn't reachable.
...
definitions:
...
  ApplicationOutputDTO:
    type: object
    required: [application_id, name, description]
    properties:
      application_id:
        type: integer
        format: int64
        description: Unique identifier representing a specific registration.
      name:
        type: string
        description: Display name of application.
      description:
        type: string
        description: Description of the application.
  ApplicationInputDTO:
    type: object
    required: [name, description, password]
    properties:
      name:
        type: string
        description: Display name of application.
      description:
        type: string
        description: Description of application.
      password:
        type: string
        description: Password of the application.
...